Migrated fragmentation strategies to separate "fragment" comodule
